fix: Timer
